# api/openapi-spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sundsvalls Kommuns öppna data",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.sundsvall.se/opendata/1.0",
      "description": "production environment"
    },
    {
      "url": "https://api-test.sundsvall.se/opendata/1.0",
      "description": "test environment"
    }
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "Returns this OpenAPI specification",
        "operationId": "getOpenAPI",
        "security": [{}],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/beaches/{id}": {
      "get": {
        "operationId": "getBeachByID",
        "description": "Get information about a single beach.",
        "parameters": [{
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "ID of the Beach to retrieve"
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "The ID of this Beach",
                          "example": "urn:ngsi-ld:Beach:se:sundsvall:facilities:553"
                        },
                        "name": {
                          "type": "string",
                          "description": "The name of this beach",
                          "example": "Bergafjärdens Havsbad"
                        },
                        "description": {
                          "type": "string",
                          "description": "A description about this beach",
                          "example": "Bergafjärden i Njurunda är ett havsbad med 800 m fin sandstrand och långgrunt, vilket gör att den passar bra för familjer med barn. Butik och servering, grillplats, minigolf, lekplats och toalett finns vid stranden. Bergafjärden är öppet maj-september och vattenprover tas. Badet ligger vid Bergafjärdens camping."
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["Point"]
                            },
                            "coordinates": {
                              "type": "array",
                              "description": "WGS84 longitude and latitude",
                              "minItems": 2,
                              "maxItems": 3,
                              "items": {
                                "type": "number"
                              },
                              "example": [17.454723, 62.266598]
                            }
                          }
                        },
                        "waterquality": {
                          "type": "array",
                          "description": "An array of water quality observations, sorted in reversed historical order",
                          "items": {
                            "type": "object",
                            "properties": {
                              "temperature": {
                                "type": "number",
                                "example": 11.7
                              },
                              "source": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://badplatsen.havochvatten.se/"
                              },
                              "dateObserved": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "weather": {
                          "type": "array",
                          "description": "An array of weather observations, sorted in reversed historical order",
                          "items": {
                            "type": "object",
                            "properties": {
                              "temperature": {
                                "type": "number",
                                "example": 16.2
                              },
                              "windDirection": {
                                "type": "number",
                                "description": "The wind direction expressed in decimal degrees compared to geographic North (measured clockwise).",
                                "minimum": 0,
                                "maximum": 360,
                                "exclusiveMaximum": true,
                                "example": 127
                              },
                              "windSpeed": {
                                "type": "number",
                                "description": "The observed wind speed in m/s.",
                                "minimum": 0,
                                "example": 6
                              },
                              "source": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://www.smhi.se/"
                              },
                              "dateObserved": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "seeAlso": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uri"
                          },
                          "example": [
                            "https://badplatsen.havochvatten.se/badplatsen/karta/#/bath/SE0712281000003475",
                            "https://www.wikidata.org/wiki/Q16498519"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/cityworks/{id}": {
      "get": {
        "operationId": "getCityWorksByID",
        "description": "Get information about a specific citywork instance",
        "parameters": [{
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "ID of the citywork to retrieve"
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "The ID of this citywork instance",
                          "example": "urn:ngsi-ld:CityWork:se:sundsvall:cityworks"
                        },
                        "description": {
                          "type": "string",
                          "description": "A description containing a reason, and contact information about this citywork instance",
                          "example": "Road closed until 2022-02-22"
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["Point"]
                            },
                            "coordinates": {
                              "type": "array",
                              "description": "WGS84 longitude and latitude",
                              "minItems": 2,
                              "maxItems": 3,
                              "items": {
                                "type": "number"
                              },
                              "example": [17.454723, 62.266598]
                            }
                          }
                        },
                        "dateModified": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/exercisetrails": {
      "get": {
        "operationId": "getExerciseTrails",
        "description": "Get information about published exercise trails",
        "parameters": [{
          "in": "query",
          "name": "fields",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["areaserved", "datelastpreparation", "description", "difficulty", "location", "paymentrequired", "source", "status"]
            }
          },
          "required": false,
          "description": "Filter the returned properties per entry"
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "The ID of this exercise trail",
                            "example": "urn:ngsi-ld:ExerciseTrail:se:sundsvall:facilities:123"
                          },
                          "name": {
                            "type": "string",
                            "description": "The name of this exercise trail",
                            "example": "Hotellslingan 5K"
                          },
                          "location": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["Point"]
                              },
                              "coordinates": {
                                "type": "array",
                                "description": "WGS84 longitude and latitude",
                                "minItems": 2,
                                "maxItems": 3,
                                "items": {
                                  "type": "number"
                                },
                                "example": [17.454723, 62.266598]
                              }
                            }
                          },
                          "categories": {
                            "type": "array",
                            "description": "List of exercise trail categories",
                            "items": {
                              "type": "string",
                              "enum": ["bike-track", "floodlit", "ice-skating", "ski-classic", "ski-skate"]
                            },
                            "example":  ["floodlit", "ski-classic"]
                          },
                          "length": {
                            "type": "number",
                            "description": "The length of this trail in kilometers",
                            "format": "float",
                            "example": 5
                          },
                          "difficulty": {
                            "type": "number",
                            "description": "A difficulty rating in the range [0.0 1.0] where 0 is easy and 1 is very hard",
                            "format": "float",
                            "example": 0.5
                          },
                          "paymentRequired": {
                            "type": "boolean",
                            "description": "A value determining if payment is required to use the trail"
                          },
                          "status": {
                            "type": "string",
                            "description": "The status of this exercise trail",
                            "enum": ["open", "closed"]
                          },
                          "dateLastPreparation": {
                            "type": "string",
                            "description": "Time of last preparation",
                            "example": "2022-10-02T03:27:18Z"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "application/geo+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "enum": ["FeatureCollection"]
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "urn:ngsi-ld:ExerciseTrail:se:sundsvall:facilities:123"
                          },
                          "type": {
                            "type": "string",
                            "enum": ["Feature"]
                          },
                          "geometry": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["LineString"]
                              },
                              "coordinates": {
                                "type": "array",
                                "items": {
                                  "type": "array",
                                  "minItems": 2,
                                  "maxItems": 3,
                                  "items": {
                                    "type": "number",
                                    "description": "WGS84 longitude and latitude"
                                  }
                                },
                                "example": [[17.454723, 62.266598], [17.454727, 62.266612]]
                              }
                            }
                          },
                          "properties": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["ExerciseTrail"]
                              },
                              "name": {
                                "type": "string",
                                "example": "Hotellslingan 5K"
                              },
                              "categories": {
                                "type": "array",
                                "description": "List of exercise trail categories",
                                "items": {
                                  "type": "string",
                                  "enum": ["bike-track", "floodlit", "ice-skating", "ski-classic", "ski-skate"]
                                },
                                "example":  ["floodlit", "ski-classic"]
                              },
                              "length": {
                                "type": "number",
                                "description": "The length of this trail in kilometers",
                                "format": "float",
                                "example": 5
                              },
                              "difficulty": {
                                "type": "number",
                                "description": "A difficulty rating in the range [0.0 1.0] where 0 is easy and 1 is very hard",
                                "format": "float",
                                "example": 0.5
                              },
                              "paymentRequired": {
                                "type": "boolean",
                                "description": "A value determining if payment is required to use the trail"
                              },
                              "status": {
                                "type": "string",
                                "description": "The status of this exercise trail",
                                "enum": ["open", "closed"]
                              },
                              "dateLastPreparation": {
                                "type": "string",
                                "description": "Time of last preparation",
                                "example": "2022-10-02T03:27:18Z"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/roadaccidents/{id}": {
      "get": {
        "operationId": "getRoadAccidentsByID",
        "description": "Get single published road accident from Trafikverket",
        "parameters": [{
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "ID of the road accident to retrieve"
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "The ID of this road accident",
                          "example": "urn:ngsi-ld:RoadAccident:se:sundsvall:facilities:553"
                        },
                        "description": {
                          "type": "string",
                          "description": "The description of this road accident",
                          "example": "Passenger car broken down in left lane."
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["Point"]
                            },
                            "coordinates": {
                              "type": "array",
                              "description": "WGS84 longitude and latitude",
                              "minItems": 2,
                              "maxItems": 3,
                              "items": {
                                "type": "number"
                              },
                              "example": [17.454723, 62.266598]
                            }
                          }
                        },
                        "accidentDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "dateCreated": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "dateModified": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "status": {
                          "type": "string",
                          "description": "The status of this road accident",
                          "example": "Ongoing"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/stratsys/publishedreports/{reportId}": {
      "get": {
        "operationId": "getStratsysReportByID",
        "description": "Get a published report by ID from stratsys.se.",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "description": "ID of the report to be retrieved",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/temperature/air": {
      "get": {
        "operationId": "getAirTemperatures",
        "description": "Get temperature values from a given sensor.<p>Available sensors are returned from the /temperature/air/sensors endpoint.",
        "parameters":[
          {
            "name": "sensor",
            "in": "query",
            "description": "ID of the temperature sensor to query.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeAt",
            "in": "query",
            "description": "Select temperatures from time",
            "example": "2021-06-01T00:00:00Z",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTimeAt",
            "in": "query",
            "description": "Select temperatures up to time",
            "example": "2021-07-01T00:00:00Z",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "options",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["aggregatedValues"]
            }
          },
          {
            "name": "aggrMethods",
            "in": "query",
            "description": "Select aggregation methods. (options MUST be set to aggregatedValues)",
            "schema": {
              "type": "string",
              "enum": ["avg", "max,min", "avg,max,min"]
            }
          },
          {
            "name": "aggrPeriodDuration",
            "in": "query",
            "description": "Select aggregation period. (options MUST be set to aggregatedValues)",
            "schema": {
              "type": "string",
              "enum": ["PT1H", "PT24H", "P7D"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}